requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
psycopg2-binary